arreglo del dashboard,uml mas reglas de negocio

+++ src/main/java/acme/datatypes/Statistics.java
@@ -18,8 +18,8 @@ public class Statistics {
 
 	Integer						count;
 	Double						average;
-	Integer						min;
-	Integer						max;
+	Double						min;
+	Double						max;
 	Double						deviation;
